pyproject.toml
[project]
name = "openapi-chat-agent"
version = "0.1.0"
description = "PoC: Turn OpenAPI into conversational agent using Google ADK"
requires-python = ">=3.12"
authors = [ { name = "Your Name" } ]
dependencies = [
	"fastapi[standard]>=0.116.1,<1.0.0",
	# uvicorn and pydantic are included with fastapi[standard]
	"pydantic-settings>=2.4,<2.6",  # Updated upper bound
	"httpx>=0.27,<0.29",  # Updated upper bound for latest releases
	"cryptography>=42.0,<44.0",  # Updated upper bound
	"google-adk>=1.12.0",  # Keep as is since it's recent
	# Database dependencies
	"sqlalchemy[asyncio]>=2.0.0,<2.1.0",
	"asyncpg>=0.29.0,<0.30.0",
	"alembic>=1.13.0,<1.14.0",
	# Authentication dependencies
	"python-jose[cryptography]>=3.3.0,<4.0.0",
	"passlib[bcrypt]>=1.7.4,<1.8.0",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.3,<8.4",
	"pytest-asyncio>=0.23,<0.25",  # Updated upper bound
]

[tool.uv]
index-url = "https://pypi.org/simple"